refactor: consolidate approval receipt ownership
