+++ tach.toml
@@ -119,6 +119,7 @@ depends_on = [
 visibility = [
     "mindroom.approval_execution",
     "mindroom.approval_manager",
+    "mindroom.approval_receipt",
     "mindroom.approval_response",
     "mindroom.approval_transport",
     "mindroom.bot",
@@ -641,7 +642,7 @@ visibility = ["mindroom.response_runner"]
 
 [[modules]]
 path = "mindroom.approval_receipt"
-depends_on = []
+depends_on = ["mindroom.event_journal"]
 visibility = [
     "mindroom.approval_execution",
     "mindroom.response_runner",
