+++ tests/test_queued_message_notify.py
@@ -14,9 +14,7 @@ import nio
 import pytest
 from agno.agent import Agent as AgnoAgent
 from agno.db.base import BaseDb, SessionType
-from agno.exceptions import ModelProviderError
 from agno.media import Image
-from agno.models.fallback import FallbackConfig, acall_model_with_fallback
 from agno.models.message import Message
 from agno.run.agent import RunCompletedEvent, RunContentEvent, RunOutput
 from agno.run.base import RunStatus
@@ -25,7 +23,7 @@ from agno.session.agent import AgentSession
 from agno.session.team import TeamSession
 from structlog.testing import capture_logs
 
-from mindroom import ai_runtime, approval_receipt
+from mindroom import ai_runtime
 from mindroom.agent_storage import create_state_storage, get_agent_session
 from mindroom.ai import _PreparedAgentRun, ai_response, stream_agent_response
 from mindroom.ai_runtime import (
@@ -3024,76 +3022,6 @@ def test_notice_hook_keeps_single_notice_at_end_and_skips_stop_after_tool_call()
     assert next_turn_messages[-1].content == QUEUED_MESSAGE_NOTICE_TEXT
     assert next_notice_context.response_turn_id != notice_context.response_turn_id
     assert _notice_count(stop_after_messages) == 0
-
-
-@pytest.mark.asyncio
-async def test_model_call_hook_appends_one_trusted_approval_receipt_after_tool_results() -> None:
-    """A resumed model must see approval provenance after its protocol-required tool result."""
-    model = RecordingModel(id="approval-receipt", provider="fake")
-    approval_receipt.install_approval_receipt_hooks(model, None)
-    messages = [
-        Message(role="system", content="base rules"),
-        Message(role="assistant", content="calling tool"),
-        Message(role="tool", content="tool result"),
-    ]
-
-    with approval_receipt.approval_receipt_context("trusted approval receipt"):
-        await model.aresponse(messages=messages)
-
-    assert [(message.role, message.content) for message in model.seen_messages] == [
-        ("system", "base rules\n\ntrusted approval receipt"),
-        ("assistant", "calling tool"),
-        ("tool", "tool result"),
-    ]
-    assert model.seen_messages[0].provider_data == {"mindroom_approval_receipt": True}
-    assert [(message.role, message.content) for message in messages] == [
-        ("system", "base rules"),
-        ("assistant", "calling tool"),
-        ("tool", "tool result"),
-        ("assistant", "ok"),
-    ]
-
-
-@pytest.mark.asyncio
-async def test_approval_receipt_reaches_fallback_model_after_primary_failure() -> None:
-    """A resumed fallback call must receive the same trusted approval provenance."""
-    primary = RecordingModel(id="primary", provider="fake")
-    fallback = RecordingModel(id="fallback", provider="fake")
-    vars(primary)["aresponse"] = AsyncMock(side_effect=ModelProviderError("primary failed"))
-    fallback_config = FallbackConfig(on_error=[fallback])
-    approval_receipt.install_approval_receipt_hooks(primary, fallback_config)
-    messages = [Message(role="system", content="base rules")]
-
-    with approval_receipt.approval_receipt_context("trusted approval receipt"):
-        response = await acall_model_with_fallback(
-            primary,
-            fallback_config,
-            messages=messages,
-        )
-
-    assert response.content == "ok"
-    assert fallback.seen_messages[0].content == "base rules\n\ntrusted approval receipt"
-    assert [(message.role, message.content) for message in messages] == [
-        ("system", "base rules"),
-    ]
-
-
-@pytest.mark.asyncio
-async def test_approval_receipt_reaches_each_concurrent_model_once() -> None:
-    """Every model participating in one resumed team run must receive the trusted receipt."""
-    first = RecordingModel(id="first-member", provider="fake")
-    second = RecordingModel(id="second-member", provider="fake")
-    approval_receipt.install_approval_receipt_hooks(first, None)
-    approval_receipt.install_approval_receipt_hooks(second, None)
-
-    with approval_receipt.approval_receipt_context("trusted approval receipt"):
-        await asyncio.gather(
-            first.aresponse(messages=[Message(role="system", content="first rules")]),
-            second.aresponse(messages=[Message(role="system", content="second rules")]),
-        )
-
-    assert first.seen_messages[0].content == "first rules\n\ntrusted approval receipt"
-    assert second.seen_messages[0].content == "second rules\n\ntrusted approval receipt"
 
 
 def test_default_queued_notice_requires_pause_handoff() -> None:
